tools/source-analyzers/BRD-static-analysis/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brd-static-analysis-py"
version = "0.1.0"
description = "Automated Business Requirements Document (BRD) Generator using Static Analysis"
authors = [
  { name = "Antigravity", email = "antigravity@example.com" },
]
dependencies = [
  "tree-sitter>=0.20.0",
  "tree-sitter-languages>=1.0.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0",
  "jinja2>=3.0.0",
  "typer>=0.9.0",
]
requires-python = ">=3.10"

[project.scripts]
brd-static-analysis-py = "brd_generator.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/brd_generator"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""
